Nav bar update; added scholarly service; added redirect from research.html; added site source link
Some papers contain a reference to 'brechtdeman.com/research.html' - this should forward to the blog. Perhaps in time it could forward to all blog posts tagged as 'research'.
The navigation bar is updated to work on mobile phones - even if it's less pretty.
URL to Github page with page source.